+++ tests/app.js
@@ -26,11 +26,13 @@ const platformSupportedModules = [];
 if (Platform.other) {
   platformSupportedModules.push('app');
   platformSupportedModules.push('functions');
+  platformSupportedModules.push('firestore');
   platformSupportedModules.push('database');
   platformSupportedModules.push('auth');
   platformSupportedModules.push('storage');
   platformSupportedModules.push('remoteConfig');
   platformSupportedModules.push('analytics');
+  platformSupportedModules.push('appCheck');
   // TODO add more modules here once they are supported.
 }
 
@@ -81,10 +83,14 @@ function loadTests(_) {
         firebase.firestore().useEmulator('localhost', 8080);
         // Firestore caches documents locally (a great feature!) and that confounds tests
         // as data from previous runs pollutes following runs until re-install the app. Clear it.
-        firebase.firestore().clearPersistence();
+        if (!Platform.other) {
+          firebase.firestore().clearPersistence();
+        }
       }
-      if (platformSupportedModules.includes('storage'))
+      if (platformSupportedModules.includes('storage')) {
         firebase.storage().useEmulator('localhost', 9199);
+        firebase.app().storage('gs://react-native-firebase-testing').useEmulator('localhost', 9199);
+      }
     });
 
     afterEach(async function afterEachTest() {